feat: [x] Create [advanceToNextPoolRound] mutation         [x] This mutation will get the correct format strategy based on Tournament Format         [x] Call the [handleNextPoolRound] method from [Format-Strategy] to handle next [Rounds]

+++ src/level/entities/levelStandings.entity.ts
@@ -47,15 +47,15 @@ export class LevelTeamStanding {
   pointsAgainst: number;
 
   @Field()
-  @Column('int')
+  @Column('decimal')
   pointsScoredByNumberOfGames: number;
 
   @Field()
-  @Column('int')
+  @Column('decimal')
   pointsAgainstByNumberOfGames: number;
 
   @Field()
-  @Column('int')
+  @Column('decimal')
   pointDiffByNumberOfGames: number;
 
   @Field()
